renaming chatmanager to extensions
chatmanager is now extensions
moved bot to root directory
adjusted bot calls in start.py
